Fix according to review.
