--- library/src/test/java/io/github/luizgrp/sectionedrecyclerviewadapter/testdoubles/stub/ParametersedSectionStub.java
@@ -1,32 +1,0 @@
-package io.github.luizgrp.sectionedrecyclerviewadapter.testdoubles.stub;
-
-import android.support.v7.widget.RecyclerView;
-import android.view.View;
-
-import io.github.luizgrp.sectionedrecyclerviewadapter.Section;
-import io.github.luizgrp.sectionedrecyclerviewadapter.SectionParameters;
-
-/**
- * A stub of Section constructed using SectionParameters
- */
-public class ParametersedSectionStub extends Section {
-
-    public ParametersedSectionStub(SectionParameters sectionParameters) {
-        super(sectionParameters);
-    }
-
-    @Override
-    public int getContentItemsTotal() {
-        return 0;
-    }
-
-    @Override
-    public RecyclerView.ViewHolder getItemViewHolder(View view) {
-        return null;
-    }
-
-    @Override
-    public void onBindItemViewHolder(RecyclerView.ViewHolder holder, int position) {
-
-    }
-}
